+++ threedigrid_builder/grid/linear.py
@@ -11,6 +11,50 @@ import pygeos
 COORD_EQUAL_ATOL = 1e-8  # the distance below which coordinates are considered equal
 
 
+def sanitize_linestrings(linestrings, points_1, points_2):
+    """Make sure that the linestrings go from points_1 to points_2.
+
+    The linestring is reversed if the distance from the linestring start to point_1
+    and v.v. is lowered.
+
+    Then, point_1 and/or point_2 are added to the linestring if the distance is larger
+    than the tolerance of 1e-8.
+
+    This function acts inplace on linestrings.
+    """
+    has_geom = np.where(pygeos.is_geometry(linestrings))[0]
+    if len(has_geom) == 0:
+        return
+
+    # reverse the geometries where necessary
+    node_start = points_1[has_geom]
+    node_end = points_2[has_geom]
+    line_start = pygeos.get_point(linestrings[has_geom], 0)
+    line_end = pygeos.get_point(linestrings[has_geom], -1)
+
+    dist_start_start = pygeos.distance(node_start, line_start)
+    dist_end_end = pygeos.distance(node_end, line_end)
+    dist_start_end = pygeos.distance(node_start, line_end)
+    dist_end_start = pygeos.distance(node_end, line_start)
+    needs_reversion = has_geom[
+        (dist_start_start > dist_start_end) & (dist_end_end > dist_end_start)
+    ]
+    if len(needs_reversion) > 0:
+        linestrings[needs_reversion] = pygeos.reverse(linestrings[needs_reversion])
+        line_start[needs_reversion] = pygeos.get_point(linestrings[needs_reversion], 0)
+        line_end[needs_reversion] = pygeos.get_point(linestrings[needs_reversion], -1)
+        dist_start_start[needs_reversion] = dist_start_end[needs_reversion]
+        dist_end_end[needs_reversion] = dist_end_start[needs_reversion]
+
+    # for the remaining geometries; add point if necessary
+    add_first = has_geom[dist_start_start > COORD_EQUAL_ATOL]
+    if len(add_first) > 0:
+        linestrings[add_first] = prepend_point(linestrings[add_first], points_1[add_first])
+    add_end = has_geom[dist_end_end > COORD_EQUAL_ATOL]
+    if len(add_end) > 0:
+        linestrings[add_end] = append_point(linestrings[add_end], points_2[add_end])
+
+
 class BaseLinear:
     content_type = None  # to be defined by subclasses
 
@@ -22,8 +66,6 @@ class BaseLinear:
         Args:
             connection_nodes (ConnectionNodes): to take the coordinates from
         """
-        has_no_geom = pygeos.is_missing(self.the_geom)
-
         # get the start and end points
         points_1 = connection_nodes.the_geom[
             connection_nodes.id_to_index(self.connection_node_start_id)
@@ -31,35 +73,16 @@ class BaseLinear:
         points_2 = connection_nodes.the_geom[
             connection_nodes.id_to_index(self.connection_node_end_id)
         ]
+        sanitize_linestrings(self.the_geom, points_1, points_2)
 
         # construct the geometries where necessary
+        has_no_geom = pygeos.is_missing(self.the_geom)
         coordinates = np.empty((np.count_nonzero(has_no_geom), 2, 2))
         coordinates[:, 0, 0] = pygeos.get_x(points_1[has_no_geom])
         coordinates[:, 0, 1] = pygeos.get_y(points_1[has_no_geom])
         coordinates[:, 1, 0] = pygeos.get_x(points_2[has_no_geom])
         coordinates[:, 1, 1] = pygeos.get_y(points_2[has_no_geom])
         self.the_geom[has_no_geom] = pygeos.linestrings(coordinates)
-
-        has_geom = np.where(~has_no_geom)[0]
-        if len(has_geom) == 0:
-            return
-
-        # reverse the geometries where necessary
-        node_start = points_1[has_geom]
-        node_end = points_2[has_geom]
-        line_start = pygeos.get_point(self.the_geom[has_geom], 0)
-        line_end = pygeos.get_point(self.the_geom[has_geom], -1)
-        needs_reversion = has_geom[
-            (
-                pygeos.distance(node_start, line_start)
-                > pygeos.distance(node_start, line_end)
-            )
-            & (
-                pygeos.distance(node_end, line_end)
-                > pygeos.distance(node_end, line_start)
-            )
-        ]
-        self.the_geom[needs_reversion] = pygeos.reverse(self.the_geom[needs_reversion])
 
     def interpolate_nodes(self, node_id_counter, global_dist_calc_points):
         """Compute nodes on each linear object with constant intervals
@@ -597,3 +620,27 @@ def segment_start_end(linestrings, segment_counts, dist_to_start):
     end_s[last_idx] = pygeos.length(linestrings)
     end_s[np.isnan(end_s)] = dist_to_start
     return start_s, end_s, counts_to_row_index(segment_counts)
+
+
+def _add_point(linestrings, points, at=0):
+    """Append or prepend points to linestrings"""
+    counts = pygeos.get_num_coordinates(linestrings)
+    coords, index = pygeos.get_coordinates(linestrings, return_index=True)
+    start, end = counts_to_ranges(counts)
+    if at == 0:
+        insert_at = start
+    elif at == -1:
+        insert_at = end
+    else:
+        raise ValueError(f"Cannot insert at {at}")
+    new_coords = np.insert(coords, insert_at, pygeos.get_coordinates(points), axis=0)
+    new_index = np.insert(index, insert_at, np.arange(len(insert_at)))
+    return pygeos.linestrings(new_coords, indices=new_index)
+
+
+def prepend_point(linestrings, points):
+    return _add_point(linestrings, points, at=0)
+
+
+def append_point(linestrings, points):
+    return _add_point(linestrings, points, at=-1)