+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog of threedigrid-builder
 1.2.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Only give node_type 4 and kcu 52/54 to manholes with a not-NULL storage area.
 
 
 1.2.1 (2022-01-27)